+++ Dfe.PrepareConversions/Dfe.PrepareConversions/Pages/TaskList/SchoolOverview/PRUPupilsAttendingGroup.cshtml
@@ -1,6 +1,6 @@
 @page "/task-list/{id:int}/confirm-school-overview/pupils-attending-group"
 @using Dfe.PrepareConversions.TagHelpers
-@model Dfe.PrepareConversions.Pages.TaskList.SchoolOverview.PRUPupilsAttendingGroup
+@model Dfe.PrepareConversions.Pages.TaskList.SchoolOverview.PruPupilsAttendingGroup
 @{
     ViewData["Title"] = (ViewData.ModelState.IsValid ? "" : "Error: ") + "Which groups of pupils attend the school?";
     Layout = "_QuestionLayout";
