+++ Dfe.PrepareConversions/Dfe.PrepareConversions/Pages/TaskList/SchoolOverview/SENNumberOfPlacesFundedFor.cshtml.cs
@@ -3,6 +3,7 @@ using Dfe.PrepareConversions.Data.Models;
 using Dfe.PrepareConversions.Data.Services;
 using Dfe.PrepareConversions.Models;
 using Dfe.PrepareConversions.Services;
+using Dfe.PrepareConversions.Utils;
 using Dfe.PrepareConversions.ViewModels;
 using Microsoft.AspNetCore.Mvc;
 using System.Threading.Tasks;
@@ -40,7 +41,7 @@ public class SENNumberOfPlacesFundedFor : BaseAcademyConversionProjectPageModel
 
       if (ModelState.IsValid)
       {
-         SetSchoolOverviewModel updatedSchoolOverview = CreateUpdateSchoolOverview(Project);
+         SetSchoolOverviewModel updatedSchoolOverview = SchoolOverviewHelper.CreateUpdateSchoolOverview(Project);
          updatedSchoolOverview.Id = id;
          updatedSchoolOverview.NumberOfPlacesFundedFor = NumberOfPlacesFundedFor;
          await _repository.SetSchoolOverview(id, updatedSchoolOverview);
@@ -50,25 +51,5 @@ public class SENNumberOfPlacesFundedFor : BaseAcademyConversionProjectPageModel
 
 
       return await base.OnGetAsync(id);
-   }
-   public static SetSchoolOverviewModel CreateUpdateSchoolOverview(ProjectViewModel projectViewModel)
-   {
-      return new SetSchoolOverviewModel(
-          projectViewModel.Id.ToInteger(0),
-          projectViewModel.PublishedAdmissionNumber,
-          projectViewModel.ViabilityIssues,
-          projectViewModel.FinancialDeficit,
-          projectViewModel.NumberOfPlacesFundedFor,
-          projectViewModel.NumberOfResidentialPlaces,
-          projectViewModel.NumberOfFundedResidentialPlaces,
-          projectViewModel.PartOfPfiScheme,
-          projectViewModel.PfiSchemeDetails,
-          projectViewModel.DistanceFromSchoolToTrustHeadquarters,
-          projectViewModel.DistanceFromSchoolToTrustHeadquartersAdditionalInformation,
-          projectViewModel.MemberOfParliamentNameAndParty,
-          projectViewModel.PupilsAttendingGroupPermanentlyExcluded,
-          projectViewModel.PupilsAttendingGroupMedicalAndHealthNeeds,
-          projectViewModel.PupilsAttendingGroupTeenageMums
-      );
    }
 }
